chore: cleaning up of vaex export command line util
