+++ packages/vaex-core/vaex/export.py
@@ -257,173 +257,3 @@ if __name__ == "__main__":
     sys.exit(main(sys.argv))
 
 
-def main(argv):
-    import argparse
-    parser = argparse.ArgumentParser(argv[0])
-    parser.add_argument('--verbose', '-v', action='count', default=0)
-    parser.add_argument('--quiet', '-q', default=False, action='store_true', help="do not output anything")
-    parser.add_argument('--list', '-l', default=False, action='store_true', help="list columns of input")
-    parser.add_argument('--progress', help="show progress (default: %(default)s)", default=True, action='store_true')
-    parser.add_argument('--no-progress', dest="progress", action='store_false')
-    parser.add_argument('--shuffle', "-s", dest="shuffle", action='store_true', default=False)
-    parser.add_argument('--sort', dest="sort", default=None)
-    parser.add_argument('--virtual', dest="virtual", action='store_true', default=False, help="Also export virtual columns")
-    parser.add_argument('--fraction', "-f", dest="fraction", type=float, default=1.0, help="fraction of input dataset to export")
-    parser.add_argument('--filter', dest="filter", default=None, help="filter to apply before exporting")
-
-    subparsers = parser.add_subparsers(help='type of input source', dest="task")
-
-    parser_soneira = subparsers.add_parser('soneira', help='create soneira peebles dataset')
-    parser_soneira.add_argument('output', help='output file')
-    parser_soneira.add_argument("columns", help="list of columns to export (or all when empty)", nargs="*")
-    parser_soneira.add_argument('--dimension', '-d', type=int, help='dimensions', default=4)
-    # parser_soneira.add_argument('--eta','-e', type=int, help='dimensions', default=3)
-    parser_soneira.add_argument('--max-level', '-m', type=int, help='dimensions', default=28)
-    parser_soneira.add_argument('--lambdas', '-l', type=int, help='lambda values for fractal', default=[1.1, 1.3, 1.6, 2.])
-
-    parser_tap = subparsers.add_parser('tap', help='use TAP (Table Access Protocol) as source')
-    parser_tap.add_argument("tap_url", help="input source or file")
-    parser_tap.add_argument("table_name", help="input source or file")
-    parser_tap.add_argument("output", help="output file (ends in .fits or .hdf5)")
-    parser_tap.add_argument("columns", help="list of columns to export (or all when empty)", nargs="*")
-
-    parser_file = subparsers.add_parser('file', help='use a file as source (e.g. .hdf5, .fits, .vot (VO table), .asc (ascii)')
-    parser_file.add_argument("input", help="input source or file, when prefixed with @ it is assumed to be a text file with a file list (one file per line)")
-    parser_file.add_argument("output", help="output file (ends in .fits or .hdf5)")
-    parser_file.add_argument("columns", help="list of columns to export (or all when empty)", nargs="*")
-
-    parser_file = subparsers.add_parser('csv', help='use a csv file as source (e.g. .hdf5, .fits, .vot (VO table), .asc (ascii)')
-    parser_file.add_argument("input", help="input source or file, when prefixed with @ it is assumed to be a text file with a file list (one file per line)")
-    parser_file.add_argument("output", help="output file (ends in .hdf5)")
-    parser_file.add_argument("columns", help="list of columns to export (or all when empty)", nargs="*")
-
-    args = parser.parse_args(argv[1:])
-
-    verbosity = ["ERROR", "WARNING", "INFO", "DEBUG"]
-    logging.getLogger("vaex").setLevel(verbosity[min(3, args.verbose)])
-    dataset = None
-    if args.task == "soneira":
-        if vaex.utils.check_memory_usage(4 * 8 * 2**args.max_level, vaex.utils.confirm_on_console):
-            if not args.quiet:
-                print("generating soneira peebles dataset...")
-            dataset = vaex.dataset_misc.SoneiraPeebles(args.dimension, 2, args.max_level, args.lambdas)
-            dataset = vaex.dataframe.DataFrameLocal(dataset)
-        else:
-            return 1
-    if args.task == "tap":
-        dataset = vaex.dataset.DatasetTap(args.tap_url, args.table_name)
-        if not args.quiet:
-            print("exporting from {tap_url} table name {table_name} to {output}".format(tap_url=args.tap_url, table_name=args.table_name, output=args.output))
-    if args.task == "csv":
-        # dataset = vaex.dataset.DatasetTap(args.tap_url, args.table_name)
-        if not args.quiet:
-            print("exporting from {input} to {output}".format(input=args.input, output=args.output))
-    if args.task == "file":
-        if args.input[0] == "@":
-            inputs = open(args.input[1:]).readlines()
-            dataset = vaex.open_many(inputs)
-        else:
-            dataset = vaex.open(args.input)
-        if not args.quiet:
-            print("exporting from {input} to {output}".format(input=args.input, output=args.output))
-
-    if dataset is None and args.task not in ["csv"]:
-        if not args.quiet:
-            print("Cannot open input")
-        return 1
-    if dataset:
-        dataset.set_active_fraction(args.fraction)
-    if args.list:
-        if not args.quiet:
-            print("columns names: " + " ".join(dataset.get_column_names()))
-    else:
-        if args.task == "csv":
-            row_count = -1  # the header does not count
-            with file(args.input) as lines:
-                for line in lines:
-                    row_count += 1
-                    # print line
-            logger.debug("row_count: %d", row_count)
-            with file(args.input) as lines:
-                line = next(lines).strip()
-                # print line
-                names = line.strip().split(",")
-                line = next(lines).strip()
-                values = line.strip().split(",")
-                numerics = []
-                for value in values:
-                    try:
-                        float(value)
-                        numerics.append(True)
-                    except:
-                        numerics.append(False)
-                names_numeric = [name for name, numeric in zip(names, numerics) if numeric]
-                print(names_numeric)
-                output = vaex.dataset_misc.Hdf5MemoryMapped.create(args.output, row_count, names_numeric)
-                Ncols = len(names)
-                cols = [output.columns[name] if numeric else None for name, numeric in zip(names, numerics)]
-
-                def copy(line, row_index):
-                    values = line.strip().split(",")
-                    for column_index in range(Ncols):
-                        if numerics[column_index]:
-                            value = float(values[column_index])
-                            cols[column_index][row_index] = value
-                row = 0
-                copy(line, row)
-                row += 1
-                progressbar = vaex.utils.progressbar(title="exporting") if args.progress else None
-                for line in lines:
-                    # print line
-                    copy(line, row)
-                    row += 1
-                    if row % 1000:
-                        progressbar.update(row / float(row_count))
-                progressbar.finish()
-                # print names
-        else:
-            if args.columns:
-                columns = args.columns
-            else:
-                columns = None
-            if columns is None:
-                columns = dataset.get_column_names(strings=True, virtual=args.virtual)
-            for column in columns:
-                if column not in dataset.get_column_names(strings=True, virtual=True):
-                    if not args.quiet:
-                        print("column %r does not exist, run with --list or -l to list all columns" % column)
-                    return 1
-
-            base, output_ext = os.path.splitext(args.output)
-            if output_ext not in [".hdf5", ".fits", ".arrow"]:
-                if not args.quiet:
-                    print("extension %s not supported, only .hdf5, .arrow and .fits are" % output_ext)
-                return 1
-
-            if not args.quiet:
-                print("exporting %d rows and %d columns" % (len(dataset), len(columns)))
-                print("columns: " + " ".join(columns))
-            progressbar = vaex.utils.progressbar(title="exporting") if args.progress else None
-
-            def update(p):
-                if progressbar:
-                    progressbar.update(p)
-                return True
-            if args.filter:
-                dataset.select(args.filter, name='export')
-                selection = 'export'
-            else:
-                selection = None
-            if output_ext == ".hdf5":
-                export_hdf5(dataset, args.output, column_names=columns, progress=update, shuffle=args.shuffle, sort=args.sort, selection=selection)
-            elif output_ext == ".arrow":
-                from vaex.arrow.export import export as export_arrow
-                export_arrow(dataset, args.output, column_names=columns, progress=update, shuffle=args.shuffle, sort=args.sort, selection=selection)
-            elif output_ext == ".fits":
-                export_fits(dataset, args.output, column_names=columns, progress=update, shuffle=args.shuffle, sort=args.sort, selection=selection)
-            if progressbar:
-                progressbar.finish()
-            if not args.quiet:
-                print("\noutput to %s" % os.path.abspath(args.output))
-            dataset.close()
-    return 0
